Rename CI_COMMIT_REF_NAME to VAR in image:docs-single job
In the Dockerfile that builds the stable versions, we have the
following variables defined:

```shell
ARG VER
ENV CI_COMMIT_REF_NAME=$VER
```

However, we never pass the `VER` variable in the `--build-arg` flags
that are used to build the Docker image, thus `CI_COMMIT_REF_NAME`
is never set, and we get a warning that it was not consumed.

This change replaces `CI_COMMIT_REF_NAME` with `VER` in the build
argument. We actually do the same thing for `image:docs-single-lunrjs`.

This worked so far because when we cut the release each month, we set
the `VAR` to be the `CI_COMMIT_REF_NAME`:
https://gitlab.com/gitlab-org/gitlab-docs/-/blob/09bed9d4f9483339748428bb3bb2603c3b2e8915/lib/tasks/release.rake#L55