Execute review job earlier
It seems like the only dependency of the `review` job are `compile_dev`
or `compile_upstream_review_app` depending on pipeline type.

If we declare that explicitly, we can execute the job right away,
allowing the pipeline to finish earlier.

Furthermore we can have a shared set of rules for `review` and
`review_stop`.